Enemies deal damage

+++ Inoculation/Assets/LevelScripts/Enemy.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Enemy : MonoBehaviour
 {
     [SerializeField]
-    private int speed = 10;
+    private float speed = 10;
     [SerializeField]
     private int health = 100;
     [SerializeField]
@@ -13,7 +13,7 @@ public class Enemy : MonoBehaviour
 
     private int currentWaypointIndex = 0;
 
-    public int getSpeed() { return speed / 10; }
+    public float getSpeed() { return speed / 10; }
     public int getHealth() { return health; }
     public int getStrength() { return strength; }
